+++ package/src/workerManager.ts
@@ -1,8 +1,8 @@
 import type { LanguageServiceDefaultsImpl } from './monaco.contribution';
 import type { KustoWorker } from './kustoWorker';
 
-import IDisposable = monaco.IDisposable;
-import Uri = monaco.Uri;
+type IDisposable = monaco.IDisposable;
+type Uri = monaco.Uri;
 
 export class WorkerManager {
     private _storedState: {
